First attempt at a Locations
This is a naive approach of just accepting what's been given as
location. Clearly this needs more work and some decisions, but this
helps with revealing the what needs to be done to make this usable.

The way I see it:

- [ ] normalising locations (unify different spelling programatically or
by enforcing it at contribution time)
- [ ] decide on a common way including multiple locations (e.g. "London;
Berlin"), document it in CONTRIBUTING and make existing data conform to
it
- [ ] generate higher level groupings by parsing country/area from
location (e.g. UK, India, USA, or maybe even Europe)